pyproject.toml
[tool.poetry]
name = "reserva-app"
version = "0.1.0"
description = ""
authors = ["Mateus <mateusc.ramos@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
bcrypt = "^4.2.0"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
